Update patch set 19
Patch Set 19:

> Patch Set 19:
> 
> > as martin, i'm disappointed that you submitted this already. it wouldn't have run away over the holidays ...
> 
> Patrick wrote:
> 
>  This decision will be made at a later point and a 'conclusion' doc will be uploaded.
> 
> which I think makes it clear that clarifications, alternate solution proposals, and so on are very welcome. It would be hard to make an informed decision without them.
> 
> Go discovered the same thing in their design doc process: at some point, keeping a single change open to track a proposed design becomes unmanageable. Once a proposal starts settling down (which I hope you'll agree had happened for this proposal), people need to ability to discuss specific modifications to it or completely different alternatives, which works best in separate changes.

Martin, it's also not clear to me what the problem is. This solution is Ben's proposal to solve the issue, it's up to Ben to decide when he thinks it's ready. The process also encourages people to submit solutions quickly and make improvements on it iteratively. So even after submit, Ben might still want to modify it. It doesn't mean that this solution was chosen yet, that decision will be done by the ESC. Other people still have time to upload alternative solutions, and they could have done this before (e.g. as follow-up change, or they could have asked to split out the use-cases if that was blocking them).

To go on, I'm interested if anyone is working on an alternative solution? If yes, do you feel like you need more time to make it ready?

I also want to point out that an alternative solution can be as easy as referencing the current solution and describing modifications/differences to it.

Patch-set: 19