Update patch set 19
Patch Set 19:

my general expectation is to fix everything at the earliest possible point in the process, because:
- postponed stuff tends to be forgotten
- committed stuff has "momentum", so *in practice* it's always harder to change things later on
- even if everything goes as planned, it results in fragmented history which is harder to research

i can see how committing early on makes sense if it's supposed to be a basis for distributed work, but i don't see this being the case here. that would be more a case for splitting off the use cases.

if managing so many comments in gerrit is a problem, then that's a test case for improvements for you right there. ;)

Patch-set: 19