Update patch set 19
Patch Set 19:

> Patch Set 19:
> 
> my general expectation is to fix everything at the earliest possible point in the process, because:
> - postponed stuff tends to be forgotten
> - committed stuff has "momentum", so *in practice* it's always harder to change things later on
> - even if everything goes as planned, it results in fragmented history which is harder to research
> 
> i can see how committing early on makes sense if it's supposed to be a basis for distributed work, but i don't see this being the case here. that would be more a case for splitting off the use cases.
> 
> if managing so many comments in gerrit is a problem, then that's a test case for improvements for you right there. ;)

I don't think that managing the number of comments is an issue here.
The problem is that reaching an agreement with everyone on all details is hard.
If an author of a solution is required to "fix everything", single reviewers can block proposals forever. E.g. if there 2 reviewers that have contrary opinions it's not possible to satisfy both. This is why the process allows to propose alternative solutions, which can be discussed separately with their pros and cons. Again, if anyone is working on an alternative solution and needs more time for this, please let us know.

As soon as an initial version of a design doc is submitted it's expected from reviewers to no longer comment on the change, but to upload improvements/changes to the design doc as Gerrit changes that can be reviewed. Please do so to ensure that your remaining issues don't get forgotten.

Patch-set: 19