Update patch set 19
Patch Set 19:

(7 comments)

> > That said, I understand where you are coming from. The problem is a bit similar to comments, which can be unresolved (strict attention). How do you make sure that resolved comments are still read (loose attention)? We don't have a great answer for that, but I would still be hesitant to add "unreplied" for loose attention, because it adds too much complexity.
> 
> i think you mean "unread".

Yes. I have added tracking "unread" changes as a "Non-Goal" to be clear that this is not covered by this design.

> As Martin, I'm disappointed that you submitted this already.

I think Martin was just disappointed that I had not split off and submitted use cases before.

> It wouldn't have run away over the holidays ...

I find it is the hardest part of software engineering to determine when something is "good enough". And I don't claim to make the best choice all the time, but my experience is that a lot of time can be wasted in pursuit of the "perfect" solution. So sometimes it is important to submit something that is not perfect such that you can start implementing earlier.

Patch-set: 19