Update patch set 19
Patch Set 19:

> Patch Set 19:
>
> Martin, it's also not clear to me what the problem is
> ...
> Other people still have time to upload alternative solutions, and they could have done this before (e.g. as follow-up change, or they could have asked to split out the use-cases if that was blocking them).

This exact request was made at least twice by Nasser on this change, see his re-pasted requests below:


> Patch Set 4:
> ...
> Only reviewing use cases for now. It'd be great to have the solution in a separate change that can be reviewed after agreement on the use cases (see https://gerrit-review.googlesource.com/Documentation/dev-design-docs.html#expectation ). That also helps get the use-cases change merged early.

> Patch Set 13:
> 
> (5 comments)
> 
> The use-cases doc feels very close to done (I just had some clarifying questions). Can you split that into its own change so it can be +2ed and submitted?


> Patch Set 19:
> 
> I think Martin was just disappointed that I had not split off and submitted use cases before.

I'm not sure that it's fair to characterize my disappointment as "just" this. That was the basis of my complaint. Once the accepted process was bypassed, and the feedback requesting the process to be followed was ignored, I felt pretty discouraged from participating more in the discussion. This discouragement then continued to grow as the review continued to grow and possibly even diverged more and more from consensus. I believe that this is result which the process/request was attempting avoid: a very large (over 150 messages) and controversial review. It feels all too easy to now conclude that consensus cannot be reached and that we might need the ESC to intervene.

Patch-set: 19