Update patch set 1
Patch Set 1:

(2 comments)

Thank you for your proposal Martin. It really helps to have alternative providing another view on this challenge. 

At the same time I have to agree with Jonathan and Ben that tacking this on as a plugin and re-/abusing the assignee for this does not seem the right way to go about such a core aspect of code review. Figuring out what a developer needs to look at next is probably one of the core tasks they need to accomplish using Gerrit.

Patch-set: 1
CC: Gerrit User 1074328 <1074328@173816e5-2b9a-37c3-8a2e-48639d4f1153>