Update patch set 3
Patch Set 3:

(2 comments)

there is a mode of using the assignee feature which hasn't been given much attention during the recent discussions: to assign a primary reviewer and leaving it at that, irrespective of whose turn it is. this would be actively broken by this solution, so it would have to be disabled/unloaded on affected sites, leaving them at square one.
otoh, i'm not sure how popular such a workflow actually is, because it makes a lot less sense than in a bug tracking system.

a formality: line wrapping throughout the text is random. you should standardize on keeping lines below 80 chars, and ideally pay some attention to making them look as little "jagged" as possible (i.e., reflow words to equalize line lengths).

Patch-set: 3