Update patch set 3
Patch Set 3: Code-Review+2

(3 comments)

The context of my review is that I am updating the conclusion doc for “Attention set” (Ia4cc7edce840).

I tried to objectively compare the proposed design in this change to the use cases. Comments that did not come out of this objective comparison are explicitly marked as my own.

I voted +2 on this change because I think we can get it merged. It is in parts an alternative to Ben’s proposal, in other parts similar. I’ll try to have that be reflected in the conclusion.

Patch-set: 3
Reviewer: Gerrit User 1026112 <1026112@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+2