Update patch set 6
Patch Set 6:

(11 comments)

> Patch Set 6:
> 
> (9 comments)
> 
> Thanks Luca and Patrick for putting so much effort into this write up. It was very well written and easy to review.
> 
> I'm excited about both solutions and would be fine with either path forward. Don't let all my inline comments distract from that 😊

Thanks, I followed up on your comments. My bottom line is that the proposal mainly differ in *how* we get to an end result, not so much in what that end results looks like (except for smaller issues). For that fact I do hope that we can wrap up discussions rather quickly and jump into implementation.

We had an ESC meeting today and discussed proposals across the following dimensions. Detailed notes will be published later for everyone to read.

1) Core vs. Plugin (not much of a concern anymore)
Unanimous conclusion: What changes a user should review when is a core feature that should not depend on plugins to work.
2) From scratch vs. changing assignee
Unanimous conclusion: We prefer a from-scratch implementation with a switch-over at a single point in time to adapting an existing feature that does not quite fit what the new feature should look like. This has the advantage of leaving 'assignee' in place for Axis for the time being and hopefully switching it off in the future when their use cases are sufficiently covered. More generally it means that current users have to adapt just once (if at all).

The discussion was opened at the summit and there was a lot of room for people to contribute and the whole thing went trough many iterations. The two proposals are not all that different, so the ESC wants to wrap this up rather quickly now and get the conclusion merged.

Patch-set: 6
Reviewer: Gerrit User 1026112 <1026112@173816e5-2b9a-37c3-8a2e-48639d4f1153>