Fix links in design docs pages
None of the links from the design docs' index pages work because:

1. They are linking to the .md files rather than the rendered html
   files.

2. They are using relative links, which don't work for nested folders,
   rather than absolute links.

Also the image links in the attention set user interface design
document don't work because they are relative links.

Fix all the permalinks to be in the correct place, and update the
documents to use absolute links. Move the attention set's images to the
/images folder, rename them, and update the document to use absolute
links.

Also remove unnecessary sidebar tags.

Change-Id: I517c523a8099010bdade9f18a707eb23fa629e4f