Update patch set 7
Patch Set 7:

(10 comments)

> Patch Set 6:
> 
> (11 comments)
> 
> > Patch Set 6:
> > 
> > (9 comments)
> > 
> > Thanks Luca and Patrick for putting so much effort into this write up. It was very well written and easy to review.
> > 
> > I'm excited about both solutions and would be fine with either path forward. Don't let all my inline comments distract from that 😊
> 
> Thanks, I followed up on your comments. My bottom line is that the proposal mainly differ in *how* we get to an end result, not so much in what that end results looks like (except for smaller issues). For that fact I do hope that we can wrap up discussions rather quickly and jump into implementation.

Yes and no. Ben's end result is also the starting point (for any users). There's no intermediate value or trying out of the sub-features that make attention set such an interesting idea. Martin's end result is hypothetically the same as Ben's, but as I note inline, might change significantly based on real usage of intermediate changes.

We've all developed features, deployed them, tried it more, and realized they didn't work the way we wanted/expected. Then we fix them, and repeat the process. It's much cheaper to do that with small iterative features than it is for a single big feature. This is why Agile was created and is so popular. This is why gerrit-review runs a build very close to the tip (I'm sure there's other reasons too, but this was the first reason I heard from Shawn). Please don't undervalue this.

> 
> We had an ESC meeting today and discussed proposals across the following dimensions. Detailed notes will be published later for everyone to read.
> 
> 1) Core vs. Plugin (not much of a concern anymore)
> Unanimous conclusion: What changes a user should review when is a core feature that should not depend on plugins to work.
> 2) From scratch vs. changing assignee
> Unanimous conclusion: We prefer a from-scratch implementation with a switch-over at a single point in time to adapting an existing feature that does not quite fit what the new feature should look like. This has the advantage of leaving 'assignee' in place for Axis for the time being and hopefully switching it off in the future when their use cases are sufficiently covered. More generally it means that current users have to adapt just once (if at all).

What does the release timeline look like for that? Assignee defaults to off in 3.2 for new sites? Forced off in 3.3? There's no plan yet for "use cases are sufficiently covered", so this feels a lot like just kicking the problem down the road.

I don't think anyone from Axis weighed in on Martin's proposal, so I really have no idea if it is less disruptive to them or not. Presumably, it would be feasible to have a config option for "# of people in the assignee set" once Martin's solution gets to that stage and perhaps that would be enough.

> 
> The discussion was opened at the summit and there was a lot of room for people to contribute and the whole thing went trough many iterations. The two proposals are not all that different, so the ESC wants to wrap this up rather quickly now and get the conclusion merged.

There was a LOT of feedback given at the summit (namely a lot of concern around UI changes and having multiple people in the attention set). Sadly, I don't see much of that feedback incorporated in the end proposal from Ben (except as Nice To Haves).

I don't think anyone has been dragging their feet on this and I understand the desire to move forward. I think it's because the solutions are so similar that we should be able to get an accurate comparison of them in the conclusion. It's feeling more like the ESC has predetermined Ben's solution the winner and this is just tidying up the paperwork. I was really hoping for more from this process.

Patch-set: 7
Reviewer: Gerrit User 1002666 <1002666@173816e5-2b9a-37c3-8a2e-48639d4f1153>