Update patch set 8
Patch Set 8:

Thanks for all your comments.

Discussing two solutions makes it sound like they are orthogonal, but after reading and rounds of discussions, I arrive at another conclusion. There is just one solution and we are discussing process and subtle differences. I think discussing this as 'Ben's proposal and 'Martin's proposal didn't do us much of a favor, because it sounds like one would loose, which I think is not the case.

So I do agree with Oswald in that further discussion cycles doesn't seem like time well spent. Parsing your initial comment ("I'm excited about both solutions and would be fine with either path forward. Don't let all my inline comments distract from that 😊") I hope that you agree as well.

The ESC voted for bootstrapping 'attention set' as standalone feature next to assignee and then (hopefully) have assignee migrated over. This vote was unanimous.

The influence that Martin's write-up had in the final outcome is certainly that there has to be more steam behind moving the 'assignee' functionality over in the end. This is reflected in this conclusion. Personally, I do think there could be plugin endpoints for customization in the end (also an influence from that write-up), but we'll be able to judge that much better once we have some feedback on how it fits into people's organizations.

There are some aspects from your comments that I want to briefly pick up:

On being 'agile':
We can't pick cherries on what 'iterative' means. "code, learn, adapt" is a contradiction to "no iterations on the API, data model or index" at least in the arguments that I’ve read here. I wouldn't label Ben's proposal as 'waterfall', there are certainly iterations to rules etc, but the project is bootstrapped with an initial revision. Agile is no excuse for intermediary states that could be perceived as half-baked. Ben/Martin have a different initial step size. We had a lengthy design discussion now, so a larger initial step size seems legitimate.

On 'feedback not being incorporated':
I think this comment is unfair to Ben. The doc had 19 patch sets and over 500 comments that came in addition to the summit discussions. This was the most open and inclusive debate I've seen in Gerrit to date. The discussion included many people, so it's natural that not every comment changed the design.

On 'Ben being the 'predetermined winner/tidying up the paperwork'
As I said in my opening statement, there were scheduling reasons why I wrote this instead of non-Google ESC members. I have high confidence that we would have arrived at the same conclusion regardless, but I've anticipated that comment to come at some point.

I do believe that we have reached the end of a challenging design process, so I am asking Luca, Alice and David to sign off on this change and get it merged - Thanks!

Patch-set: 8
Reviewer: Gerrit User 1026112 <1026112@173816e5-2b9a-37c3-8a2e-48639d4f1153>