Update patch set 3
Abandoned

Patrick ran some numbers and found that assuming over 300 endpoints and around 5 permission tests for each endpoint, performing the suggestion described in the document would result in a very high number of tests (5! for each endpoint on average). Anyway, this is abandoned since there are higher priorities right now, and also there is an easier alternative: We can run a script to check for coverage in Gerrit (run all tests in Gerrit and see which lines of code are never run). This idea is also better since we check for other issues, and not just security issues.

Patch-set: 3
Status: abandoned
Tag: autogenerated:gerrit:abandon