Update patch set 1
Patch Set 1:

Can we revert this?  These are basically 5 people at Google that are championing the use of Gerrit for linux. I want to do work in this space in 2020, and  even got some moderate funding for it, but I think we should see more adoption before we spin this as a "Gerrit is being used".

Patch-set: 1
Reviewer: Gerrit User 1024147 <1024147@173816e5-2b9a-37c3-8a2e-48639d4f1153>