Code menu: Add a plugins page to list all plugins
Provide a Gerrit plugins compatibility matrix on the homepage, this way:

- Add python script to generate the plugins page from markdown template.
- Explain the purpose of the page/matrix and how it works, in that page.
- Add publishing instructions on how to update this page, for latest CI.

In a nutshell, maintainers are expected to use plugins.py to re-generate
plugins.md. That script re-generates the latter from template.md and the
latest data from [1], [2] and [3]:

[1] https://gerrit-review.googlesource.com
[2] https://gerrit-ci.gerritforge.com
[3] https://gerrit.googlesource.com

The hereby added python script is inspired from David's [4], used for
ESC [5] to initiate a plugins survey between 2018.11.30 and 2019.05.30
last year. The script also implements ideas such as [6] from Matthias.

[4] https://gist.github.com/dpursehouse/66e1bd861efce4631ca1966b85ca0226
[5] https://www.gerritcodereview.com/members.html#engineering-steering-committee
[6] https://www.cloudbees.com/blog/taming-jenkins-json-api-depth-and-tree

More sophisticated styling and formatting could be considered as
follow-up changes. Same goes for further automation to update the page
dynamically from CI. The latter ([2]) or similar could even host that
information in the future.

Feature: Issue 11400
Helped-by: David Pursehouse <dpursehouse@collab.net>
Helped-by: Matthias Sohn <matthias.sohn@sap.com>
Change-Id: I2c81d2822e2a27908bb10abba2e3df60e3def55b