Restructure plugin generation files
Move the pipenv config and flake8 settings to the project root.

Move the plugins.py and template.md to a new tools/ folder. This
separates the tools from the markdown sources that are generated
and published.

Update the instructions accordingly.

Note that if the pipenv has previously been installed, it will no
longer work with the moved files and will need to be reinstalled.

Change-Id: Ib52812916140d7d0c8e47001faf50f3ebe16f310