Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> maybe we should rename template.md to plugins-template.md since it's now located in a generic tools folder

I'm planning to rework the script to use jinja2 for the template handling; the file will get renamed as part of that.

BTW do you have CR+2 on this repo?  If not, we should fix that.

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>