Plugin list: improve finding plugin maintainers
So far we used a naming convention to find the owner group for plugins.
This isn't really accurate, hence get the owner groups using the
projects endpoint of the REST API. This now also handles plugins having
multiple owner groups correctly.

If the parent project is "Public-Plugins" add the core maintainers
to the list of plugin maintainers.

Change-Id: Ie48759045d96275e1c7fb3983a65d13348f1e5fb