Update patch set 18
Patch Set 18: Code-Review+1

(1 comment)

I took an AI from the ESC meeting to look at the doc. Thanks for writing things up in great detail.

Adding an instance ID/name to events seems cheap and doesn't add much tech debt. Receivers or setups that don't care about the ID can just not send it or disregard it when processing events.

So overall, I think this is of no harm and if it solves problems in your setup we should just go with it.

From a system design perspective, I am not convinced that solving the problem of not wanting to receive events a server dispatched itself should be solved in Gerrit. I would expect the queue backend to either:
- Be configured to never dispatch an event to the sender
- Allow for ACK'ing an event that one dispatched right after sending it: When you dispatch a message, you should get a unique ID. Use that to ACK the message for the sender instantly.

Patch-set: 18
Reviewer: Gerrit User 1026112 <1026112@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+1