Update patch set 23
Patch Set 23: Code-Review-1

Maybe it would help for me to clarify that I was never one to ask for this design doc process. In general my experiences with it so far have been negative.

If someone were to upload a change to Core today that added the ability for plugins to contribute data to events, with the explanation that this was going to be used for solving problems like "I want to add event source information from the HA/multi-site plugin", and there were good details about risks and impacts in the commit message, I would expect that change to be rapidly approved by all maintainers. If that's all you need, why bother with this design doc in the first place?

If there's an insistence on having a design doc, then I think we should hold those docs to a high standard. They should tackle more than what we could do with a single change and commit message, otherwise it's a waste of everyone's time. The wording around "Use-Cases" in https://gerrit-review.googlesource.com/Documentation/dev-design-docs.html#structure was intentionally chosen. "The interactions between a user and a system to attain particular goals." I do not think that criteria is met in this document as of yet.

Patch-set: 23
Reviewer: Gerrit User 1002666 <1002666@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1