Update patch set 3
Patch Set 3: -Code-Review

Sorry, I forgot about this change, and in the meantime Luca has updated the 2.16.18 release notes separately.

This change should then be rebased on top of the master branch, and the 3.0 and 3.1 release note modifications should be synced with the text that was added for 2.16.18

Note that I also have a follow-up change under review that rewords it a bit, so maybe better to wait until that one has been submitted.

Patch-set: 3
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=0