Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I understood that the goal of this plan was to have every RC exactly one week apart, and bounded by non-weekend days (work days).

Yes, this is more of an exception I would say.

> There should also be RC4 to catch up on RC3 leftovers. Unless RC3 has close to no delta to deliver since RC2?

I believe there is WIP that *could* go into RC3 if we delay that by a couple of days.
I won't instead open up so much in new features in RC4, as the RC3 to RC4 is more for fixes only.

Patch-set: 1