Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > We should at least have the change for updating the replication plugin documentation: I believe it would be really useful to have feedback on this new exciting feature.
> 
> There is a doc entry in the config doc for replication.distributionInterval. Are you looking for something that documents how to correctly setup a site to make use of the distribution?

Yes, I can point to https://gerrit.googlesource.com/plugins/replication/+/refs/heads/master/src/main/resources/Documentation/config.md but the formatting isn't great and it doesn't give the high-level view.

It would be nice to have a high-level view of how to setup a replication clustering properly and with a nicer formatting. What about just amending the design and put a real-life example of the implementation and configuration?

Patch-set: 1