Gerrit 3.2: amend release notes with more configuration changes
Amend existing Gerrit 3.2 release notes by adding additional changes
that have been made to the Gerrit configuration and the associated
behaviours in Gerrit.

Details were initially missed at the time of the release.

Move also the native packages changes to the top so
that can be more prominent and noticed by the Gerrit admins,
because of the double upgrade of Java and Operating System.

Change-Id: I14a4c1b9878207a7c94145ccf6e30fdef248838a