Move Dave Borowitz to former maintainers
Dave moved to another project 1 year ago and is no longer active as a
Gerrit maintainer. I've asked him whether it's okay to remove him from
the maintainer list, so that the it accurately reflects the active
maintainers, and he agreed.

Once this change is submitted, I will follow-up with removing Dave from
the maintainers group.

As soon as Dave finds time to make regular contributions to Gerrit
again, we can add him back as a maintainer.

Bug: Issue 12750
Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I4adcb0ff99c75ee768372c0973750faa4f587306