Update PolyGerrit screenshots on the homepage
Update Gerrit screenshots on the homepage with the
new PolyGerrit UX.

NOTE: The real-estate for the screenshots is very narrow
and the images are looking very small and difficult to
read. Revamping the homepage layout is out of scope for
this change, and it should be addressed as a follow-up.

Bug: Issue 11391
Change-Id: I8090bcda47a018d7e4d7a0d3e8cc15e7e982d0b7