Update PolyGerrit screenshots on the homepage
Update Gerrit screenshots on the homepage with the
new PolyGerrit UX.

NOTE: The real-estate for the screenshots is very narrow
and the images are looking very small and difficult to
read. It is out of scope of this change the revamp of the
homepage layout and sizing, which it should be addressed
in a separate change.

Bug: Issue 11391
Change-Id: I8090bcda47a018d7e4d7a0d3e8cc15e7e982d0b7