Upload thoughts on reworking the event system
During the hackathon in May 2019, I gave a presentation on the
issues of Gerrit's current event system and how we could rework
it to a better system. Here are my extensive notes on this
matter.

I currently don't have any intention to drive this redesign.
I'm just uploading this so that others have access to my
notes/thoughts, too. Do whatever you want with them.

Change-Id: Ic5f844681554f7abaac5b35a98440a93a8165f0a