Upload thoughts on reworking the event system
During the hackathon in May 2019, I gave a presentation on the
issues of Gerrit's current event system and how we could rework
it to a better system. Here are my extensive notes on this
matter.

I currently don't have any intention to drive this redesign.
I'm just uploading this so that others have access to my
notes/thoughts, too.

Change-Id: Ic29adb80ee0be67b0288a45d8fedfdd3c459cf8b