Expand 3.3 release candidate notes with more plugin information
Add information that is either manually collected (for specific topics),
or manually added as release_noter excluded it. Such exclusions, albeit
rather seldom, are being tuned accordingly, as the script gets improved;
cf. Issue 13515 for that matter.

The removed plugin (core, submodule) changes are caused by Issue 13517,
which is yet to be addressed. Remove other non-noteworthy content
spotted while working on this change.

Bug: Issue 13483
Change-Id: I8abba57ad49f35d6cc6df3d0b2d051bc323197bb