Remove GerritForge Enterprise Support contract details
Follow-up of the Change Ie06a035ad to address the feedback raised
on duplicated information between the GerritForge's support website
and the Gerrit Code Review one.

Point to the GerritForge's web-site for extra clarity on what is
provided by the company as part of their Enterprise Support contract
for their clients.

Change-Id: Iec117e9aea647417701637ef18a43fc97bff2bef