Clarify the EOL for Gerrit v2.16
During the discussion [1] of the meeting minutes of the 1st
of June ESC the term "community support" is relevant only
for GerritForge but not for the rest of the Gerrit users
and contributors.

The decision during the ESC was only about dropping the
"special support" for Gerrit v2.16 and did not imply any
change in the Gerrit support process: remove the term
"community support" to avoid any confusion.

[1] https://groups.google.com/g/repo-discuss/c/uL-sPJyHz9M

Change-Id: Iefe9d3bdfa1a1b2e4be3d500668f8715621f7974