List update-ref command for Gerrit 3.3 downgrade
To downgrade All-Projects.git ref/meta/version, one can use a shell
oneliner to bring the version down to 183:

git update-ref refs/meta/version $(echo -n 183|git hash-object --stdin)

Mention it in the doc to save one time to have to look at the man page.

A oneliner would be nice for All-Users.git to revert schema-184
migration, I have just updated the refs/meta/group-names reference to
the previous one but it is not ideal.

Change-Id: If7323be62c32a9d39f8cb2718ee1561d639fcc9b