Declare EOL for v3.2 and post-EOL support for v2.16
Gerrit v2.16 is officially EOL *without exceptions*;
however, members of the community having to deal with
Gerrit v2.16 may be available to help and support on
the repo-discuss mailing list.
Add the extra section in the support page to publicize
this option.

Gerrit v3.2 is EOL and allows easy and zero-downtime
migration to either v3.3, v3.4 or v3.5.

Change-Id: I4f7a8a1d248ebd023ed8e31063aa261782611769