[RFC] Proposed post-EOL support for Gerrit v2.16
Gerrit v2.16 will be officially EOL *without exceptions*
as soon as v3.5 will be released.
However, members of the community having to deal with
Gerrit v2.16 would be still willing to help and support
on the repo-discuss mailing list.

Add the extra section in the support page to publicize
this option.

NOTE: This change is out for discussion and is NOT supposed
to be merged until v3.5 is out and v2.16 is officially EOL
without exceptions.

Change-Id: I4f7a8a1d248ebd023ed8e31063aa261782611769