Push review solution 2: Describe an unsolved issue of this design
As commented here [1] creating a push review for a commit that contains
a Change-Id in the commit message likely causes confusion to users &
tooling, because the Change-Id in the commit message doesn't match the
Change-Id of the push review change.

We don't know how to solve this yet, or if this is solvable at all.
Adding this as an unsolved issue, as it's important to be aware of this
when making the conclusion and chosing a solution.

[1] https://gerrit-review.googlesource.com/c/homepage/+/335395/3/pages/design-docs/push-reviews/solution-google-2.md#27

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ia50e2fa0063ffb2acfa67f9c72cadb5fdc5cd03e