Extend the v3.6.0 release plan by one more week
The Change 336054 has not been finalised in time for the release
of v3.6.0 and the P1 Issue 15919 is tracing the current status of
the problems with the review label when migrating from v3.4 or v3.5
to v3.6.0.

The extra week should allow to finalise the migration and assure
a smoother upgrade experience to the Gerrit community.

Change-Id: I984c6922669dcedcd622e4665d86afa3ce332314