Add more details on how to import project changes between servers
Gerrit v3.7 has introduced the ability to import changes coming
from projects and servers having different ServerId and different
account ids.

Add more details on the release notes with the pointers to the
Gerrit v3.7 documentation.
Also add Luca's talk pointer with more details to the import
process, also with references from the release notes.

With the slides, the recording of the talk and the live-demo
the Gerrit administrator should have enough resources to perform
the import operation.

Change-Id: I79c2c1ccf8ad85d46c84488c5c1b20358769d9b1