Revert "Describe the new version fields on Gerrit issues"
This reverts commit e93ff9e1b0c5e83fc5ea46e2815f565f9107efe7.

Reason for revert:

The new fields didn't work as intended and are currently not used.

The fields are not visible to regular users (due to [1]) and hence
cannot be populated by issue reporters. That's why the version field
is no longer mandatory. The TargetMilestone field has not been used
since 3.6.

[1] https://bugs.chromium.org/p/monorail/issues/detail?id=10106

Change-Id: I39837fdfbf9c84f64fce2093237b38d9d09a7fc9