Remove links to Monorail from historic ESC minutes
The Gerrit issues have been migrated from Monorail into the Gerrit
Tracker. Due to this links to Monorail are no longer working, and if
they would they would no longer show the issues at that time. Hence just
drop the link (rather than pointing to the issues of the ESC component
in the new location).

Change-Id: I4446de64fde7157f1bf2d7a5cc059a7b4807ecd2
Signed-off-by: Edwin Kempin <ekempin@google.com>