Move Change 360646 into the breaking changes section
The introduction of Change 360646 caused a radical change
in the way Gerrit threads are handling the updates of the ref-database.
Move the description of the changes introduced into the "Breaking
Changes" section and add a specific note to warn about the
consequences of the new behaviour and associated risks.

The Gerrit administrator can therefore make a conscious decision
about leaving the performance feature enabled or disabled, based
on the trade-offs between speed and consistency.

Bug: Issue 309098227
Change-Id: I83d7dc34563e381a8b6b7bb82ec820063cc0e16b